pyproject.toml
[project]
name = "ichatbio-globi-agent"
version = "1"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk~=0.2.1",
    "openai~=1.79.0",
    "pydantic~=2.11.4",
    "instructor~=1.11.3",
    "dotenv~=0.9.9",
    "pytest~=8.3.5",
    "pytest-asyncio~=1.0.0a1",
    "uvicorn~=0.34.3",
    "pytest_httpx~=0.35.0"
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
log_cli = true
log_cli_level = "INFO"